feat: refactored extension functions
